Extract `evaluateStyle` and `emToPx` helper functions
