+++ lib/em-to-px.js
@@ -1,0 +1,2 @@
+const emToPx = (value, base) => Number.parseFloat(value) * base;
+module.exports = emToPx;
